renamed SoundPlayer to FlutterSoundPlayer for better backward compatibility.
